refactor: move components from pages to component folder, remove deleteModal in teams
