+++ src/modules/Organization/Pages/Teams.tsx
@@ -1,36 +1,21 @@
 import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import {
-    Dialog,
-    DialogContent,
-    DialogFooter,
-    DialogHeader,
-    DialogTitle,
-    DialogDescription,
-} from "@/components/ui/dialog";
+import {Dialog, DialogContent, DialogFooter, DialogHeader, DialogTitle, DialogDescription,} from "@/components/ui/dialog";
 import { toast } from "@/components/ui/use-toast";
 import { getErrorMessage } from "~/utils/errorHandler.ts";
 import { TeamResponse } from "@/constants/types/teamTypes";
-import {deleteTeam, getTeam} from "@/services/teamService";
+import {deleteTeam, getTeams} from "@/services/teamService";
 import {Pencil, Trash} from "lucide-react";
 import { Button } from "@/components/ui/button";
-import { Alert, AlertDescription } from "@/components/ui/alert";
 import { Card } from "@/components/ui/card";
-import {
-    Table,
-    TableBody,
-    TableCell,
-    TableHead,
-    TableHeader,
-    TableRow,
-} from "@/components/ui/table";
+import {Table, TableBody, TableCell, TableHead, TableHeader, TableRow,} from "@/components/ui/table";
 import { Pagination } from '../../../core/components';
 import {usePageTitle} from "@/contexts/PageTitleContext.tsx";
+import {DeleteModal} from "@/modules/Organization/Components/DeleteTeamDialog.tsx";
 
 export default function Teams() {
     const [teamList, setTeamList] = useState<TeamResponse[]>([]);
     const [selectedTeam, setSelectedTeam] = useState<TeamResponse | null>(null);
-    const [errorMessage, setErrorMessage] = useState<string | null>(null);
     const [isProcessing, setIsProcessing] = useState<boolean>(false);
     const navigate = useNavigate();
     const [currentPage, setCurrentPage] = useState<number>(1);
@@ -45,7 +30,7 @@ export default function Teams() {
     }, [setTitle, setChildren, navigate]);
 
     useEffect(() => {
-        getTeam()
+        getTeams()
             .then((response: TeamResponse[]) => {
                 setTeamList(response);
             })
@@ -95,54 +80,46 @@ export default function Teams() {
     );
 
     return (
-        <>
-            {errorMessage && (
-                <Alert>
-                    <AlertDescription>{errorMessage}</AlertDescription>
-                </Alert>
-            )}
+        <main className="flex flex-1 flex-col  p-4">
+            <Card className="flex flex-1 flex-col rounded-lg border border-dashed shadow-sm p-4 ">
+                <Table>
+                    <TableHeader>
+                        <TableRow>
+                            <TableHead>Name</TableHead>
+                            <TableHead className='text-right pr-8'>Actions</TableHead>
+                        </TableRow>
+                    </TableHeader>
+                    <TableBody>
+                        {paginatedTeamList.map((t) => (
+                            <TeamRowItem
+                                key={t.id}
+                                t={t}
+                                setSelectedTeam={setSelectedTeam}
+                                isProcessing={isProcessing}
+                                updateTeam={updateTeam}
+                            />
+                        ))}
+                    </TableBody>
+                </Table>
 
-            <main className="flex flex-1 flex-col  p-4">
-                <Card className="flex flex-1 flex-col rounded-lg border border-dashed shadow-sm p-4 ">
-                    <Table>
-                        <TableHeader>
-                            <TableRow>
-                                <TableHead>Name</TableHead>
-                                <TableHead className='text-right pr-8'>Actions</TableHead>
-                            </TableRow>
-                        </TableHeader>
-                        <TableBody>
-                            {paginatedTeamList.map((t) => (
-                                <TeamRowItem
-                                    key={t.id}
-                                    t={t}
-                                    setSelectedTeam={setSelectedTeam}
-                                    isProcessing={isProcessing}
-                                    updateTeam={updateTeam}
-                                />
-                            ))}
-                        </TableBody>
-                    </Table>
+                {selectedTeam && (
+                    <DeleteModal
+                        team={selectedTeam}
+                        handleReject={() => setSelectedTeam(null)}
+                        handleAccept={handleRemoveTeam}
+                    />
+                )}
 
-                    {selectedTeam && (
-                        <DeleteModal
-                            team={selectedTeam}
-                            handleReject={() => setSelectedTeam(null)}
-                            handleAccept={handleRemoveTeam}
-                        />
-                    )}
-
-                    {teamList.length > recordsPerPage && (
-                        <Pagination
-                            pageSize={recordsPerPage}
-                            pageNumber={currentPage}
-                            setPageNumber={setCurrentPage}
-                            totalContents={teamList.length}
-                        />
-                    )}
-                </Card>
-            </main>
-        </>
+                {teamList.length > recordsPerPage && (
+                    <Pagination
+                        pageSize={recordsPerPage}
+                        pageNumber={currentPage}
+                        setPageNumber={setCurrentPage}
+                        totalContents={teamList.length}
+                    />
+                )}
+            </Card>
+        </main>
     );
 }
 
@@ -180,34 +157,5 @@ function TeamRowItem({ t, isProcessing, setSelectedTeam, updateTeam }: TeamItemP
                 </div>
             </TableCell>
         </TableRow>
-    );
-}
-
-type DeleteModalProps = {
-    handleAccept: () => void;
-    handleReject: () => void;
-    team: TeamResponse;
-};
-
-function DeleteModal({ handleAccept, handleReject, team }: DeleteModalProps) {
-    return (
-        <Dialog open={true} onOpenChange={handleReject}>
-            <DialogContent className="sm:max-w-md">
-                <DialogHeader>
-                    <DialogTitle>Remove Team</DialogTitle>
-                </DialogHeader>
-                <DialogDescription>
-                    Are you sure you want to remove the {team.name} team?
-                </DialogDescription>
-                <DialogFooter className="flex justify-center">
-                    <Button onClick={handleReject} variant="outline">
-                        No
-                    </Button>
-                    <Button onClick={handleAccept} variant="destructive" className="ml-4">
-                        Yes
-                    </Button>
-                </DialogFooter>
-            </DialogContent>
-        </Dialog>
     );
 }
