+++ src/modules/Organization/index.ts
@@ -3,7 +3,7 @@ import LeavePolicy from '../Leaves/Pages/LeavePolicy.tsx'
 import OfficialHolidays from './Pages/OfficialHolidays'
 import Teams from './Pages/Teams'
 import CreateTeam from './Pages/CreateTeam'
-import UpdateTeam from './Pages/UpdateTeamDialog.tsx'
+import UpdateTeam from './Components/UpdateTeamDialog.tsx'
 import ImportHolidays from './Pages/ImportHolidays'
 import LeavesPage from '../Leaves/Pages/LeavesPage.tsx'
